--- a/foyer-storage/src/engine.rs
+++ b/foyer-storage/src/engine.rs
@@ -32,7 +32,6 @@ use crate::{
     storage::{
         either::{Either, EitherConfig, Selection, Selector},
         noop::Noop,
-        runtime::{Runtime, RuntimeStoreConfig},
     },
     DeviceStats, IoBytes, Storage,
 };
@@ -93,46 +92,34 @@ where
     }
 }
 
-enum StoreFuture<F1, F2, F3, F4, F5, F6, F7> {
+enum StoreFuture<F1, F2, F3, F4> {
     Noop(F1),
     Large(F2),
-    LargeRuntime(F3),
-    Small(F4),
-    SmallRuntime(F5),
-    Combined(F6),
-    CombinedRuntime(F7),
+    Small(F3),
+    Combined(F4),
 }
 
-impl<F1, F2, F3, F4, F5, F6, F7> StoreFuture<F1, F2, F3, F4, F5, F6, F7> {
+impl<F1, F2, F3, F4> StoreFuture<F1, F2, F3, F4> {
     // TODO(MrCroxx): use `expect` after `lint_reasons` is stable.
     #[allow(clippy::type_complexity)]
-    pub fn as_pin_mut(
-        self: Pin<&mut Self>,
-    ) -> StoreFuture<Pin<&mut F1>, Pin<&mut F2>, Pin<&mut F3>, Pin<&mut F4>, Pin<&mut F5>, Pin<&mut F6>, Pin<&mut F7>>
-    {
+    pub fn as_pin_mut(self: Pin<&mut Self>) -> StoreFuture<Pin<&mut F1>, Pin<&mut F2>, Pin<&mut F3>, Pin<&mut F4>> {
         unsafe {
             match *Pin::get_unchecked_mut(self) {
                 StoreFuture::Noop(ref mut inner) => StoreFuture::Noop(Pin::new_unchecked(inner)),
                 StoreFuture::Large(ref mut inner) => StoreFuture::Large(Pin::new_unchecked(inner)),
-                StoreFuture::LargeRuntime(ref mut inner) => StoreFuture::LargeRuntime(Pin::new_unchecked(inner)),
                 StoreFuture::Small(ref mut inner) => StoreFuture::Small(Pin::new_unchecked(inner)),
-                StoreFuture::SmallRuntime(ref mut inner) => StoreFuture::SmallRuntime(Pin::new_unchecked(inner)),
                 StoreFuture::Combined(ref mut inner) => StoreFuture::Combined(Pin::new_unchecked(inner)),
-                StoreFuture::CombinedRuntime(ref mut inner) => StoreFuture::CombinedRuntime(Pin::new_unchecked(inner)),
             }
         }
     }
 }
 
-impl<F1, F2, F3, F4, F5, F6, F7> Future for StoreFuture<F1, F2, F3, F4, F5, F6, F7>
+impl<F1, F2, F3, F4> Future for StoreFuture<F1, F2, F3, F4>
 where
     F1: Future,
     F2: Future<Output = F1::Output>,
     F3: Future<Output = F1::Output>,
     F4: Future<Output = F1::Output>,
-    F5: Future<Output = F1::Output>,
-    F6: Future<Output = F1::Output>,
-    F7: Future<Output = F1::Output>,
 {
     type Output = F1::Output;
 
@@ -140,11 +127,8 @@ where
         match self.as_pin_mut() {
             StoreFuture::Noop(future) => future.poll(cx),
             StoreFuture::Large(future) => future.poll(cx),
-            StoreFuture::LargeRuntime(future) => future.poll(cx),
             StoreFuture::Small(future) => future.poll(cx),
-            StoreFuture::SmallRuntime(future) => future.poll(cx),
             StoreFuture::Combined(future) => future.poll(cx),
-            StoreFuture::CombinedRuntime(future) => future.poll(cx),
         }
     }
 }
@@ -159,15 +143,8 @@ where
 {
     Noop,
     Large(GenericLargeStorageConfig<K, V, S>),
-    LargeRuntime(RuntimeStoreConfig<GenericLargeStorage<K, V, S>>),
     Small(GenericSmallStorageConfig<K, V, S>),
-    SmallRuntime(RuntimeStoreConfig<GenericSmallStorage<K, V, S>>),
     Combined(EitherConfig<K, V, S, GenericSmallStorage<K, V, S>, GenericLargeStorage<K, V, S>, SizeSelector<K, V, S>>),
-    CombinedRuntime(
-        RuntimeStoreConfig<
-            Either<K, V, S, GenericSmallStorage<K, V, S>, GenericLargeStorage<K, V, S>, SizeSelector<K, V, S>>,
-        >,
-    ),
 }
 
 impl<K, V, S> Debug for EngineConfig<K, V, S>
@@ -180,11 +157,8 @@ where
         match self {
             Self::Noop => write!(f, "Noop"),
             Self::Large(config) => f.debug_tuple("Large").field(config).finish(),
-            Self::LargeRuntime(config) => f.debug_tuple("LargeRuntime").field(config).finish(),
             Self::Small(config) => f.debug_tuple("Small").field(config).finish(),
-            Self::SmallRuntime(config) => f.debug_tuple("SmallRuntime").field(config).finish(),
             Self::Combined(config) => f.debug_tuple("Combined").field(config).finish(),
-            Self::CombinedRuntime(config) => f.debug_tuple("CombinedRuntime").field(config).finish(),
         }
     }
 }
@@ -201,18 +175,10 @@ where
     Noop(Noop<K, V, S>),
     /// Large object disk cache.
     Large(GenericLargeStorage<K, V, S>),
-    /// Large object disk cache with a dedicated runtime.
-    LargeRuntime(Runtime<GenericLargeStorage<K, V, S>>),
     /// Small object disk cache.
     Small(GenericSmallStorage<K, V, S>),
-    /// Small object disk cache with a dedicated runtime.
-    SmallRuntime(Runtime<GenericSmallStorage<K, V, S>>),
     /// Combined large and small object disk cache.
     Combined(Either<K, V, S, GenericSmallStorage<K, V, S>, GenericLargeStorage<K, V, S>, SizeSelector<K, V, S>>),
-    /// Combined large and small object disk cache with a dedicated runtime.
-    CombinedRuntime(
-        Runtime<Either<K, V, S, GenericSmallStorage<K, V, S>, GenericLargeStorage<K, V, S>, SizeSelector<K, V, S>>>,
-    ),
 }
 
 impl<K, V, S> Debug for Engine<K, V, S>
@@ -225,11 +191,8 @@ where
         match self {
             Self::Noop(storage) => f.debug_tuple("Noop").field(storage).finish(),
             Self::Large(storage) => f.debug_tuple("Large").field(storage).finish(),
-            Self::LargeRuntime(storage) => f.debug_tuple("LargeRuntime").field(storage).finish(),
             Self::Small(storage) => f.debug_tuple("Small").field(storage).finish(),
-            Self::SmallRuntime(storage) => f.debug_tuple("SmallRuntime").field(storage).finish(),
             Self::Combined(storage) => f.debug_tuple("Combined").field(storage).finish(),
-            Self::CombinedRuntime(storage) => f.debug_tuple("CombinedRuntime").field(storage).finish(),
         }
     }
 }
@@ -244,11 +207,8 @@ where
         match self {
             Self::Noop(storage) => Self::Noop(storage.clone()),
             Self::Large(storage) => Self::Large(storage.clone()),
-            Self::LargeRuntime(storage) => Self::LargeRuntime(storage.clone()),
             Self::Small(storage) => Self::Small(storage.clone()),
-            Self::SmallRuntime(storage) => Self::SmallRuntime(storage.clone()),
             Self::Combined(storage) => Self::Combined(storage.clone()),
-            Self::CombinedRuntime(storage) => Self::CombinedRuntime(storage.clone()),
         }
     }
 }
@@ -268,11 +228,8 @@ where
         match config {
             EngineConfig::Noop => Ok(Self::Noop(Noop::open(()).await?)),
             EngineConfig::Large(config) => Ok(Self::Large(GenericLargeStorage::open(config).await?)),
-            EngineConfig::LargeRuntime(config) => Ok(Self::LargeRuntime(Runtime::open(config).await?)),
             EngineConfig::Small(config) => Ok(Self::Small(GenericSmallStorage::open(config).await?)),
-            EngineConfig::SmallRuntime(config) => Ok(Self::SmallRuntime(Runtime::open(config).await?)),
             EngineConfig::Combined(config) => Ok(Self::Combined(Either::open(config).await?)),
-            EngineConfig::CombinedRuntime(config) => Ok(Self::CombinedRuntime(Runtime::open(config).await?)),
         }
     }
 
@@ -280,11 +237,8 @@ where
         match self {
             Engine::Noop(storage) => storage.close().await,
             Engine::Large(storage) => storage.close().await,
-            Engine::LargeRuntime(storage) => storage.close().await,
             Engine::Small(storage) => storage.close().await,
-            Engine::SmallRuntime(storage) => storage.close().await,
             Engine::Combined(storage) => storage.close().await,
-            Engine::CombinedRuntime(storage) => storage.close().await,
         }
     }
 
@@ -292,11 +246,8 @@ where
         match self {
             Engine::Noop(storage) => storage.enqueue(entry, buffer, info),
             Engine::Large(storage) => storage.enqueue(entry, buffer, info),
-            Engine::LargeRuntime(storage) => storage.enqueue(entry, buffer, info),
             Engine::Small(storage) => storage.enqueue(entry, buffer, info),
-            Engine::SmallRuntime(storage) => storage.enqueue(entry, buffer, info),
             Engine::Combined(storage) => storage.enqueue(entry, buffer, info),
-            Engine::CombinedRuntime(storage) => storage.enqueue(entry, buffer, info),
         }
     }
 
@@ -304,11 +255,8 @@ where
         match self {
             Engine::Noop(storage) => StoreFuture::Noop(storage.load(hash)),
             Engine::Large(storage) => StoreFuture::Large(storage.load(hash)),
-            Engine::LargeRuntime(storage) => StoreFuture::LargeRuntime(storage.load(hash)),
             Engine::Small(storage) => StoreFuture::Small(storage.load(hash)),
-            Engine::SmallRuntime(storage) => StoreFuture::SmallRuntime(storage.load(hash)),
             Engine::Combined(storage) => StoreFuture::Combined(storage.load(hash)),
-            Engine::CombinedRuntime(storage) => StoreFuture::CombinedRuntime(storage.load(hash)),
         }
     }
 
@@ -316,11 +264,8 @@ where
         match self {
             Engine::Noop(storage) => storage.delete(hash),
             Engine::Large(storage) => storage.delete(hash),
-            Engine::LargeRuntime(storage) => storage.delete(hash),
             Engine::Small(storage) => storage.delete(hash),
-            Engine::SmallRuntime(storage) => storage.delete(hash),
             Engine::Combined(storage) => storage.delete(hash),
-            Engine::CombinedRuntime(storage) => storage.delete(hash),
         }
     }
 
@@ -328,11 +273,8 @@ where
         match self {
             Engine::Noop(storage) => storage.may_contains(hash),
             Engine::Large(storage) => storage.may_contains(hash),
-            Engine::LargeRuntime(storage) => storage.may_contains(hash),
             Engine::Small(storage) => storage.may_contains(hash),
-            Engine::SmallRuntime(storage) => storage.may_contains(hash),
             Engine::Combined(storage) => storage.may_contains(hash),
-            Engine::CombinedRuntime(storage) => storage.may_contains(hash),
         }
     }
 
@@ -340,11 +282,8 @@ where
         match self {
             Engine::Noop(storage) => storage.destroy().await,
             Engine::Large(storage) => storage.destroy().await,
-            Engine::LargeRuntime(storage) => storage.destroy().await,
             Engine::Small(storage) => storage.destroy().await,
-            Engine::SmallRuntime(storage) => storage.destroy().await,
             Engine::Combined(storage) => storage.destroy().await,
-            Engine::CombinedRuntime(storage) => storage.destroy().await,
         }
     }
 
@@ -352,11 +291,8 @@ where
         match self {
             Engine::Noop(storage) => storage.stats(),
             Engine::Large(storage) => storage.stats(),
-            Engine::LargeRuntime(storage) => storage.stats(),
             Engine::Small(storage) => storage.stats(),
-            Engine::SmallRuntime(storage) => storage.stats(),
             Engine::Combined(storage) => storage.stats(),
-            Engine::CombinedRuntime(storage) => storage.stats(),
         }
     }
 
@@ -364,11 +300,8 @@ where
         match self {
             Engine::Noop(storage) => StoreFuture::Noop(storage.wait()),
             Engine::Large(storage) => StoreFuture::Large(storage.wait()),
-            Engine::LargeRuntime(storage) => StoreFuture::LargeRuntime(storage.wait()),
             Engine::Small(storage) => StoreFuture::Small(storage.wait()),
-            Engine::SmallRuntime(storage) => StoreFuture::SmallRuntime(storage.wait()),
             Engine::Combined(storage) => StoreFuture::Combined(storage.wait()),
-            Engine::CombinedRuntime(storage) => StoreFuture::CombinedRuntime(storage.wait()),
         }
     }
 
@@ -376,11 +309,8 @@ where
         match self {
             Engine::Noop(storage) => storage.runtime(),
             Engine::Large(storage) => storage.runtime(),
-            Engine::LargeRuntime(storage) => storage.runtime(),
             Engine::Small(storage) => storage.runtime(),
-            Engine::SmallRuntime(storage) => storage.runtime(),
             Engine::Combined(storage) => storage.runtime(),
-            Engine::CombinedRuntime(storage) => storage.runtime(),
         }
     }
 }