--- a/foyer-storage/src/storage/mod.rs
+++ b/foyer-storage/src/storage/mod.rs
@@ -14,7 +14,6 @@
 
 pub mod either;
 pub mod noop;
-pub mod runtime;
 
 use std::{fmt::Debug, future::Future, sync::Arc};