@MrCroxx 2024-08-13
# Benchmark

Run on my colleague's computer:

```bash
rm -rf /home/croxx/data/ && cargo build --release && RUST_BACKTRACE=1 RUST_LOG=info ./target/release/foyer-bench --dir /home/croxx/data --mem 64 --disk 102400 --file-size 64 --get-range 10000 --flushers 4 --reclaimers 4 --time 60 --writers 256 --w-rate 4 --admission-rate-limit 500 --readers 1024 --r-rate 1 --runtime --compression none --distribution zipf --distribution-zipf-s 0.8 --metrics
```

## main

```plain
Total:
disk total iops: 5009.8
disk total throughput: 584.6 MiB/s
disk read iops: 1262.8
disk read throughput: 83.9 MiB/s
disk write iops: 3747.1
disk write throughput: 500.8 MiB/s
insert iops: 14262.1/s
insert throughput: 891.5 MiB/s
insert lat p50: 4us
insert lat p90: 6us
insert lat p99: 13us
insert lat p999: 71us
insert lat p9999: 301us
insert lat p99999: 723us
insert lat pmax: 1055us
get iops: 17338.3/s
get miss: 8.64%
get throughput: 1.0 GiB/s
get hit lat p50: 4us
get hit lat p90: 1079us
get hit lat p99: 2767us
get hit lat p999: 3871us
get hit lat p9999: 5887us
get hit lat p99999: 7327us
get hit lat pmax: 8703us
get miss lat p50: 232us
get miss lat p90: 971us
get miss lat p99: 2063us
get miss lat p999: 3071us
get miss lat p9999: 3583us
get miss lat p99999: 4015us
get miss lat pmax: 4015us
```

## PR

```plain
Total:
disk total iops: 4999.0
disk total throughput: 585.1 MiB/s
disk read iops: 1269.3
disk read throughput: 84.3 MiB/s
disk write iops: 3729.7
disk write throughput: 500.8 MiB/s
insert iops: 14274.6/s
insert throughput: 892.3 MiB/s
insert lat p50: 4us
insert lat p90: 7us
insert lat p99: 14us
insert lat p999: 67us
insert lat p9999: 285us
insert lat p99999: 627us
insert lat pmax: 991us
get iops: 17373.5/s
get miss: 8.82%
get throughput: 1.0 GiB/s
get hit lat p50: 4us
get hit lat p90: 939us
get hit lat p99: 2639us
get hit lat p999: 4799us
get hit lat p9999: 6207us
get hit lat p99999: 7679us
get hit lat pmax: 9215us
get miss lat p50: 196us
get miss lat p90: 879us
get miss lat p99: 1807us
get miss lat p999: 2751us
get miss lat p9999: 3247us
get miss lat p99999: 3855us
get miss lat pmax: 3855us
```